requirements.txt
flask==2.1.0
jinja2==3.1.1
werkzeug==2.1.1
itsdangerous==2.1.2
flask-cors==4.0.0
python-dotenv==0.20.0
requests==2.27.1
